+++ docs/changelog.rst
@@ -63,7 +63,7 @@ Changed
 Fixed
 -----
 - Bug in `raster.transform` with lazy coordinates. (#801)
-- Bug in `workflows.mesh.mesh2d_from_rasterdataset` with multi-dimensional coordinates. (#843)
+- Bug in `model.processes.mesh.mesh2d_from_rasterdataset` with multi-dimensional coordinates. (#843)
 - Bug in `MeshModel.get_mesh` after xugrid update to 0.9.0. (#848)
 - Bug in `raster.clip_bbox` when bbox doesn't overlap with raster. (#860)
 - Allow for string format in zoom_level path, e.g. `{zoom_level:02d}` (#851)
@@ -73,7 +73,7 @@ Fixed
 Deprecated
 ----------
 - The `DataCatalog.from_archive` method is deprecated. Use `DataCatalog.from_yml` with the root pointing to the archive instead. (#849)
-- Bug in `workflows.mesh.mesh2d_from_rasterdataset` with multi-dimensional coordinates. (#843)
+- Bug in `model.processes.mesh.mesh2d_from_rasterdataset` with multi-dimensional coordinates. (#843)
 - Bug in `MeshModel.get_mesh` after xugrid update to 0.9.0. (#848)
 - Bug in `raster.clip_bbox` when bbox doesn't overlap with raster. (#860)
 - Allow for string format in zoom_level path, e.g. `{zoom_level:02d}` (#851)
